+++ packages/imperative/src/imperative/src/config/cmd/edit/edit.definition.ts
@@ -24,8 +24,7 @@ export const editDefinition: ICommandDefinition = {
     summary: "edit config files",
     description: `Edit an existing config file in your system's default text editor.\n\n` +
         `In a graphical environment, the application associated with JSON files will be launched. ` +
-        `In a command-line environment, vi will be launched. To override the command-line editor used, specify it in ` +
-        `the ${ImperativeConfig.instance.loadedConfig.envVariablePrefix}_EDITOR environment variable.`,
+        `In a command-line environment, vi will be launched. To override the command-line editor used, specify it with the --editor option.`,
     options: [
         {
             name: "global-config",
@@ -40,7 +39,15 @@ export const editDefinition: ICommandDefinition = {
             aliases: ["uc"],
             type: "boolean",
             defaultValue: false
-        }
+        },
+        {
+            name: "editor",
+            description: `Editor that overrides the default editor for this file type. Set the option to the editor's executable file location ` +
+                `or the program's name: ie "--editor notepad"`,
+            aliases: ["ed"],
+            type: "string",
+            required: false
+        },
     ],
     examples: [
         {